+++ src/ui/SettingsDialog.h
@@ -8,6 +8,7 @@
 class QCheckBox;
 class QLabel;
 class QLineEdit;
+class QComboBox;
 class QSpinBox;
 
 class SettingsDialog : public QDialog
@@ -25,6 +26,7 @@ private:
     QString m_configFilePath;
 
     QLineEdit *m_configPathEdit = nullptr;
+    QComboBox *m_appThemeCombo = nullptr;
     QLineEdit *m_terminalFontFamilyEdit = nullptr;
     QSpinBox *m_terminalFontSizeSpin = nullptr;
     QCheckBox *m_configBackupsCheck = nullptr;